Moving some types around
